feat: use g11n-langneg for deciding format data
We need to use proper language negotiation when
when loading data with multiple locale possibilities.

In order to do that, it was necessary to find a way
to transpile a commonjs-based, imported from node_modules,
and deliver it to the browser, concatenated with the rest
of the code in the file "ts.js". Browserify was used to
do that, because it plays well with our legacy Grunt, and
because it gives a lot less overhead than rewriting the
Grunt tasks to some other heavyweight bundler like Webpack
or Rollup or [insert new bundler here].

Browserify is an odd choice for 2020, but it is
the best cost-benefit for the moment. And we get now the
possibility of importing other npm deps into tradeshift-ui.